Update 05_scale_ingredients.py
This code should now work.  I shifted the code that asks for the ingredient name so that it is AFTER the if block in lines 62-72.  This means that if the users have more than 2 ingredients and they enter the exit code, they are not asked to put in an ingredient name (the 'break' command breaks out of the loop).  If the user tries to exit with too few items in the list, the program will give an error and the code now uses the 'continue' command to restart the loop (line 69-70)